Full working Dat-DNS thing
My horrible frankenstein version of the Dat DNS feature in Beaker Browser works!!!

Basically now it badges if a Dat equivalent of the open site exists and provides the option to go to it in the extension pop up. Simple.

It sadly doesn't do the pretty URLs thing and I don't think I'll ever deliver that.